Drop backwards-compatibility shim
records_mover/s3_url.py existed because there was a direct import from
an internal use a while back and I wanted to refactor this into the
then-new 'url' package.  I searched internal github and I'm no longer
finding it, so I think we're good!  Even if we weren't, we could add a
shim in the internal project pointing to the right location; if
someone were importing this they'd be pointing to that project anyway!