Allow IAM username suffix to be added to S3 scratch URL (#69)
This PR retires the scratch-s3-bucket script and introduces the ability to use an IAM username as part of the scratch bucket URL with appropriate config files in place. I'll work to get those set up so we can release this and have it work for our internal use.

I also moved the scratch bucket calculation over to BaseCreds, and renamed the functions to include the word 'default'. Your AWS IAM username depends on which AWS creds you are using. We always use the default AWS creds when loading/unloading on Redshift. That might not be true in the future, and once we provide more flexibility there, we'll want to extend this code to take in the appropriate credentials name for the AWS creds.