Add `pip freeze` results to sphinx_requirements.txt
By upgrading the airflow version, pip was unable to be a real dependency
manager and figure out what dependencies it needed to install ???
I tried to give it some hints about what versions of some other
packages it needed, but it didn't help. So I clobbered the whole thing
with the result of `pip freeze` and made a few manual changes necessary
to get things to build.