Fix Table to Table integration tests (#177)
* uncomment tbl2tbl tests

* Add more info to table_validator assertion error regarding column types

Now print the actual column types as well as the expected types.

* Change TIMESTAMPTZ to TIMESTAMP WITH TIME ZONE for some postgres tests

* Add expected column types for redshift2postgresql integration test

This is needed because redshift doesn't support the exact same types as
postgres (no TIME type) so some columns have to be converted.

* Update python version to 3.9 for tbl2tbl integration test

* Add .circleci/config.yml

* Update config.yml

* Updated config.yml

* Change local_infile option from True to 'ON'

SQLAlchemy requires query options to be strings. This might be failing due to
a new version of SQLAlchemy. From reading MySQL documentation here:

https://dev.mysql.com/doc/refman/8.0/en/server-system-variables.html#sysvar_local_infile

I think the right string to pass is 'ON'

* Change ON to '1'

sqlalchemy tries to parse this field as an int(), so I can only guess that
'1' is ON and '0' is 'OFF'

* Update expected TIMESTAMP column types

* Revert "Update expected TIMESTAMP column types"

This reverts commit 8d905e5cc98c3c90383e2d78dcf9ccfe3fad5e01.

---------

Co-authored-by: Bruno Castro-Karney <bruno.castrokarney@bluelabs.com>
Co-authored-by: Bruno Castro-Karney <bruno.castrokarney@gmail.com>